# SAGA/init_scenario_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления сценариями",
        "title": "Init Scenario API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Возвращает статус OK если сервис работает",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Проверка здоровья сервиса",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scenario/init": {
            "post": {
                "description": "Создает новый сценарий и отправляет событие в Kafka",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scenario"
                ],
                "summary": "Инициализировать сценарий",
                "parameters": [
                    {
                        "description": "Данные для создания сценария",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InitScenarioRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.InitScenarioResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.InitScenarioRequest": {
            "type": "object",
            "required": [
                "camera_id"
            ],
            "properties": {
                "camera_id": {
                    "type": "integer"
                }
            }
        },
        "dto.InitScenarioResponse": {
            "type": "object",
            "properties": {
                "scenario_uuid": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}